requirements.txt
python-telegram-bot==20.8
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-mock==3.14.0
python-dotenv==1.0.1
zabbix-api==0.5.4
selenium==4.18.1
webdriver-manager==4.0.1
matplotlib==3.8.3
requests==2.31.0
aiohttp==3.9.1
asyncio-mqtt==0.16.1 